+++ CHANGELOG.md
@@ -378,3 +378,8 @@
 - [Patch v5.2.7] Update dependency versions in requirements.txt
 - QA: pytest -q passed (207 tests)
 
+
+### 2025-07-11
+- [Patch v5.2.7] Catch UnboundLocalError in PREPARE_TRAIN_DATA
+- New/Updated unit tests added for src.main
+- QA: pytest -q passed (199 tests)